backend/requirements.txt
dj-database-url>=3.1.2
django>=6.0.3
django-cors-headers>=4.9.0
google-api-python-client>=2.194.0
google-auth>=2.49.2
Pillow>=11.3.0
psycopg[binary]>=3.3.3
requests>=2.33.1
whitenoise>=6.12.0
